requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pillow==10.4.0
piexif==1.1.3
boto3==1.35.20
botocore==1.35.20
python-dotenv==1.0.1
firebase-admin==6.6.0
httpx==0.27.2
websockets==12.0
Wand==0.6.13
opencv-python-headless==4.10.0.84
numpy==1.26.4
jinja2==3.1.4
rembg[cpu]==2.0.53
onnxruntime==1.18.1